+++ readBvecsFromTwix.m
@@ -1,18 +1,18 @@
-function readBvecsFromTwix(infolder, infile, outfolder, T)
+function readBvecsFromTwix(infile, outfolder, T)
 %readBvecsFromTwix Read b-vectors and b-values from a Siemens twix raw data file
 %
 % Requires:
 %   mapVBVD from https://github.com/pehses/mapVBVD
 %
 % Examples:
-%     readBvecsFromTwix('input_twix_folder/', 'input_twix_file.dat', 'output_folder/');
+%     readBvecsFromTwix('input_twix_folder/input_twix_file.dat', 'output_folder/');
 %   will read in the twix file `input_twix_folder/input_twix_file.dat` and write out
 %   the *b*-vectors and *b*-values of the last acquired line in each repetition to
 %   `output_folder/input_twix_file.bvec` and `output_folder/input_twix_file.bval`,
 %   respectively. It will also output the nominal *b*-values to 
 %   `output_folder/input_twix_file_nominal.bval`.
 %
-%     readBvecsFromTwix('input_twix_folder/', 'input_twix_file.dat', 'output_folder/', diag([1,-1,1]));
+%     readBvecsFromTwix('input_twix_folder/input_twix_file.dat', 'output_folder/', diag([1,-1,1]));
 %   will do the same as the previous example, but apply the transformation matrix
 %   diag([1,-1,1]) to the b-vectors before saving them. This is useful to 
 %   e.g. apply the transformation from DICOM to NIfTI space.
@@ -27,7 +27,7 @@ if ~exist('T','var')
 end
 
 %% read in data
-dat=mapVBVD(fullfile(infolder,infile));
+dat=mapVBVD(infile);
 
 nRep=dat.image.NRep;